Termdet: Implement switch-taskpool interface
When a thread finds a task it checks whether the task is from a
different taskpool as the last task and switches taskpools, if
necessary. If no task is found, the taskpool is set to NULL.
A taskpool switch causes the local task counter to be flushed
to the global counter. A thread with an active taskpool
has a pending action on the taskpool, which is released once the
thread switches away from the taskpool.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>