Implement hash table API providing a key handle during lock
The new function parsec_hash_table_lock_bucket_handle provides a key handle
that contains the key and the hash. It can then be used with other
parsec_hash_table_*_handle() functions to insert, remove, and find
items, and to unlock the bucket, all without having to rehash the key.
Rehashing a key is potentially expensive as it invokes user-provided code
through a function pointer.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>